+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Improve error messages on missing directives in module-info [#205](https://github.com/nbbrd/java-service-util/issues/205)
+
 ## [1.5.2] - 2022-10-28
 
 ### Fixed